// website/src/components/types.ts
export interface ImageData {
  url: string
  width: number
  height: number
  description?: string
  caption?: string
  author?: string
  focusPoint?: {x: number; y: number}
  format: string

  ogURL: string
  smallTeaserURL: string
  mediumTeaserURL: string
  largeURL: string
}

export interface Author {
  id: string
  url: string
  slug: string
  name: string
  image: ImageData
  bio?: RichTextBlock[]
  links?: any[]
}
// ...
export enum BlockType {
  Foo = 'foo',
  Bar = 'bar',

  // Content
  TitleImage = 'titleImage',
  Title = 'title',
  RichText = 'RichTextBlock',
  Gallery = 'gallery',
  Teaser = 'teaser',
  Embed = 'embed',
  Quote = 'quote',
  Image = 'image',
  Listicle = 'listicle',
  PeerPageBreak = 'linkPageBreak',

  // Layout
  Grid = 'grid'
}

export interface BaseBlock<T extends BlockType, V> {
  type: T
  key: string
  value: V
}

// Content Blocks
// --------------
export enum HeaderType {
  Default = 'default',
  Breaking = 'breaking'
}
// ...
export type RichTextBlock = BaseBlock<BlockType.RichText, Node[]>